extract shared type spec emission into spec module
